Move project details logic to project settings folder
