+++ client/src/containers/ProjectSettings/logic/state.ts
@@ -13,7 +13,6 @@ export const initialState: ProjectState = {
 		icon: '',
 		sprints: [],
 		boards: [],
-		icon: '',
 		defaultAssignee: undefined,
 		lead: {
 			id: '',
